Add script for devs to add own WFH whitelisting
